Created component for Assignee and used it elsewhere
