+++ mvn_setup/tasks-parent/tasks-ui/src/core/TaskEdit/TaskEditDialog.tsx
@@ -9,7 +9,6 @@ import Events from './TaskEvents';
 import Burger from '@the-wrench-io/react-burger';
 import TaskClient from '@taskclient';
 
-
 const Left: React.FC<{}> = () => {
 
   return (
@@ -57,7 +56,7 @@ const Header: React.FC<{ onClose: () => void }> = ({ onClose }) => {
       <Grid item md={6} lg={6} alignSelf='center'>
         <Stack spacing={2} direction='row'>
           <Fields.Status />
-          <Fields.Assignee />
+          <Fields.TaskAssignee />
           <Fields.Priority />
           <Fields.Options />
         </Stack>
